+++ client/dive-common/review/types.ts
@@ -41,6 +41,12 @@ export interface ReviewFrameGeometry {
 export interface ReviewFrameRef extends ReviewFrameGeometry {
   frame: number;
   bounds: RectBounds;
+  /**
+   * The track has no detection on this frame in this camera; `bounds` is
+   * interpolated from its neighbours so the chip can still be cropped there,
+   * and no box is drawn.
+   */
+  missing?: boolean;
 }
 
 /**
@@ -71,6 +77,18 @@ export interface ReviewItem {
     value: unknown;
     scope: 'track' | 'detection';
   };
+}
+
+/**
+ * One grid entry: a track shown once per camera it appears in. Single
+ * camera datasets have one item per entry; the cameras of a multicamera
+ * dataset share the entry, with their items' frames aligned.
+ */
+export interface ReviewEntry {
+  key: string;
+  items: ReviewItem[];
+  /** Camera name per item; empty strings for single-camera entries. */
+  labels: string[];
 }
 
 /** Grid presentation settings, persisted per browser. */
